requirements_fixed.txt
absl-py==2.4.0
accelerate==1.12.0
addict==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyascii==0.3.3
anyio==4.12.1
attrs==25.4.0
audioread==3.1.0
av==16.1.0
azure-core==1.38.1
azure-storage-blob==12.28.0
babel==2.18.0
bangla==0.0.5
basicsr==1.4.2
blinker==1.9.0
blis==1.3.3
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
confection==0.1.5
contourpy==1.3.3
coqpit==0.0.17
cryptography==46.0.5
ctranslate2==4.7.1
cycler==0.12.1
cymem==2.0.13
Cython==3.2.4
dateparser==1.1.8
decorator==5.2.1
docopt==0.6.2
edge-tts==7.2.7
einops==0.8.2
encodec==0.1.1
face-alignment==1.4.1
facexlib==0.3.0
fastapi==0.129.0
faster-whisper==1.2.1
filelock==3.20.0
filterpy==1.4.5
Flask==3.1.2
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
future==1.0.0
g2pkk==0.1.2
gfpgan==1.3.8
grpcio==1.78.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
h11==0.16.0
hangul-romanize==0.1.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
ImageIO==2.37.2
inflect==7.5.0
iniconfig==2.3.0
isodate==0.7.2
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
joblib==1.5.3
jsonlines==1.2.0
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.46.0
lmdb==1.7.5
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.1
murmurhash==1.0.15
networkx==3.6.1
nltk==3.9.2
num2words==0.5.14
numba==0.63.1
numpy==2.3.5
ollama==0.6.1
onnxruntime==1.24.1
openai-whisper==20250625
opencv-python==4.13.0.92
packaging @ file:///opt/miniconda3/conda-bld/packaging_1761049079066/work
pandas==1.5.3
pillow==12.1.1
platformdirs==4.9.2
pluggy==1.6.0
pooch==1.9.0
preshed==3.0.12
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
pypinyin==0.55.0
pysbd==0.3.4
pytest==9.0.2
pytest-asyncio==1.3.0
python-crfsuite==0.9.12
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
sounddevice==0.5.5
soundfile==0.13.1
soxr==1.0.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.2
starlette==0.52.1
SudachiDict-core==20260116
SudachiPy==0.6.10
sympy==1.13.1
tabulate==0.9.0
tb-nightly==2.21.0a20251023
tensorboard==2.20.0
tensorboard-data-server==0.7.2
thinc==8.3.10
threadpoolctl==3.6.0
tifffile==2026.2.16
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.5.0
torchaudio==2.5.0
torchvision==0.20.0
tqdm==4.67.3
trainer==0.0.36
transformers==5.2.0
TTS==0.22.0
typeguard==4.5.0
typer==0.24.0
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
umap-learn==0.5.11
Unidecode==1.4.0
urllib3==2.6.3
uvicorn==0.41.0
uvloop==0.22.1
wasabi==1.1.3
watchfiles==1.1.1
weasel==0.4.3
websockets==16.0
Werkzeug==3.1.5
wrapt==2.1.1
yacs==0.1.8
yapf==0.43.0
yarl==1.22.0
